requirement.txt
pytest
appium-python-client
selenium
PyYaml
allure-pytest